backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic-settings==2.2.1
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
apscheduler==3.10.4
redis==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
